Tidy up past grants search
